pyproject.toml
[project]
name = "hvadvilduhelst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Kasper Junge", email = "kasperjuunge@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "datasets>=3.6.0",
    "huggingface-hub>=0.33.0",
    "jupyter>=1.1.1",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
]

[project.scripts]
hvadvilduhelst = "hvadvilduhelst:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
